+++ op-acceptance-tests/tests/sync/elsync/gap_clp2p/sync_test.go
@@ -1,4 +1,4 @@
-package elsync
+package gap_clp2p
 
 import (
 	"testing"
